Move specific API to EntityConnection
