+++ packages/psss-server/src/routes/country/FetchCountries.ts
@@ -4,15 +4,12 @@
  */
 
 import { getSortByKey } from '@tupaia/utils';
-import { PSSS_ENTITY, PSSS_PERMISSION_GROUP } from '../../constants';
+import { PSSS_PERMISSION_GROUP } from '../../constants';
 import { Route } from '../Route';
 
 export class FetchCountries extends Route {
   async buildResponse() {
-    const countries = await this.entityConnection.fetchDescendants(PSSS_ENTITY, {
-      fields: ['id', 'code', 'name'],
-      filter: { type: 'country' },
-    });
+    const countries = await this.entityConnection.fetchCountries();
     const { accessPolicy } = await this.getSession();
     const allowedEntities = accessPolicy.getEntitiesAllowed(PSSS_PERMISSION_GROUP);
